cumulus_test_runtime
Type Alias AllPallets

pub type AllPallets = AllPalletsWithSystem;

👎Deprecated: The type definition has changed from representing all pallets excluding system, in reversed order to become the representation of all pallets including system pallet in regular order. For this reason it is encouraged to use explicitly one of AllPalletsWithSystem, AllPalletsWithoutSystem, AllPalletsWithSystemReversed, AllPalletsWithoutSystemReversed. Note that the type frame_executive::Executive expects one of AllPalletsWithSystem , AllPalletsWithSystemReversed, AllPalletsReversedWithSystemFirst. More details in https://github.com/paritytech/substrate/pull/10043

All pallets included in the runtime as a nested tuple of types.