Crate cumulus_test_runtime

Modules

api

wasm_spec_version_incremented

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockHashCount

ConstU8

Const getter for a basic type.

ConstantMultiplier

Implementor of [WeightToFee] that uses a constant multiplier.

CreationFee

ExistentialDeposit

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1.0 and adding 0.

GenesisConfig

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

MaxReserves

MinimumPeriod

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParachainId

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

TestOnRuntimeUpgrade

TransactionByteFee

TransferFee

Version

Weight

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

DispatchClass

A generalized group of dispatch types.

OriginCaller

RuntimeCall

RuntimeEvent

SudoCall

Contains one variant per dispatchable that can be called by an extrinsic.

TestPalletCall

Contains one variant per dispatchable that can be called by an extrinsic.

TimestampCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

DAYS

EPOCH_DURATION_IN_BLOCKS

HOURS

MILLISECS_PER_BLOCK

MINUTES

PRIMARY_PROBABILITY

SLOT_DURATION

TEST_RUNTIME_UPGRADE_KEY

Some key that we set in genesis and only read in TestOnRuntimeUpgrade to ensure that [OnRuntimeUpgrade] works as expected.

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_REF_TIME_PER_SECOND

Traits

BuildStorage

Complex storage builder stuff.

GetLastTimestamp

Randomness

A trait that is able to provide randomness.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Balance

Balance of an account.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CheckedExtrinsic

Extrinsic type that has already been checked.

Executive

Executive: handles dispatch to the various modules.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Index

Index of a transaction in the chain.

NodeBlock

Opaque block type.

ParachainSystem

ParachainSystemConfig

SessionHandlers

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

TestPallet

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.