Crate cumulus_test_runtime

Modules

api

elastic_scaling

elastic_scaling_500ms

elastic_scaling_multi_block_slot

elastic_scaling_mvp

relay_parent_offset

sync_backing

wasm_spec_version_incremented

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

BlockExecutionWeight

Weight of executing an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockHashCount

ConstU8

Const getter for a basic type.

ConstantMultiplier

Implementor of WeightToFee that uses a constant multiplier.

CreationFee

ExistentialDeposit

ExtrinsicBaseWeight

Weight of executing a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1.0 and adding 0.

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

MaxReserves

MinimumPeriod

Offset

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PotId

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

TestOnRuntimeUpgrade

TransactionByteFee

TransferFee

Version

Weight

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

DispatchClass

A generalized group of dispatch types.

GluttonCall

Contains a variant per dispatchable extrinsic that this pallet has.

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

SudoCall

Contains a variant per dispatchable extrinsic that this pallet has.

TestPalletCall

A special pallet that exposes dispatchables that are only useful for testing. Contains a variant per dispatchable extrinsic that this pallet has.

TimestampCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

BLOCK_PROCESSING_VELOCITY

DAYS

EPOCH_DURATION_IN_BLOCKS

HOURS

MINUTES

PARACHAIN_ID

The para-id used in this runtime.

PRIMARY_PROBABILITY

RUNTIME_API_VERSIONS

SLOT_DURATION

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

WEIGHT_REF_TIME_PER_SECOND

Traits

BuildStorage

Complex storage builder stuff.

GetLastTimestamp

Randomness

A trait that is able to provide randomness.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

build_state

Build GenesisConfig from a JSON blob not using any defaults and store it in the storage. For more info refer to sp_genesis_builder::GenesisBuilder::build_state.

get_preset

Get the default GenesisConfig as a JSON blob if name is None.

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Balance

Balance of an account.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

Executive

Executive: handles dispatch to the various modules.

Glutton

GluttonConfig

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

NodeBlock

Opaque block type.

Nonce

Index of a transaction in the chain.

Now

The current time for the current block.

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

SessionHandlers

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedPayload

The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

TestPallet

TestPalletConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

TxExtension

The extension to the basic transaction logic.

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

WeightReclaim